test file init
